% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jea_theme.R
\name{jea_theme}
\alias{jea_theme}
\title{Custom Theme for Jordan's Package}
\usage{
jea_theme()
}
\description{
Based in theme_bw() this theme function elevates the simplicity of the theme, centers labels, and places boxes around the legend.
}
\examples{
df <- data.frame( row_id = 1:10, category_a = c("A", "A", "B", "B", "C", "C", "D", "D", "E", "E"), category_b = c("1", "2", "3", "1", "1", "2", "2", "3", "1", "1"))
ggplot(df, aes(x = category_a, fill = category_b)) +geom_bar(postion = "fill") + jea_theme() + scale_fill_manual(values = jea_colors)
}
