% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propByCat.R
\name{propByCat}
\alias{propByCat}
\title{Proportion by Categorical Attribute}
\usage{
propByCat(df, cat1, cat2)
}
\arguments{
\item{df}{A Data Frame}

\item{cat1}{a string of the column name for the first category, the one the user should want to cat2 out of}

\item{cat2}{a string of the column name for the second category, the user should want to know this variable by cat1}
}
\value{
a data frame with each cat1 and cat2 and their proportions
}
\description{
Create a new dataframe with the proportion of category two for each group in category one
}
\examples{
df <- data.frame( row_id = 1:10, category_a = c("A", "A", "B", "B", "C", "C", "D", "D", "E", "E"), category_b = c("1", "2", "3", "1", "1", "2", "2", "3", "1", "1"))
new_df <- propByCat(df, "category_a", "category_b")
print(new_df)
}
