% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jea_theme.R
\name{jea_theme}
\alias{jea_theme}
\title{Custom Theme for Jordan's Package}
\usage{
jea_theme()
}
\description{
Based in theme_bw() this theme function elevates the simplicity of the theme, centers labels, and places boxes around the legend.
}
